requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
flask==3.0.2
flask-cors==4.0.0

# Database and data handling
numpy==1.26.4

# ML and Analysis dependencies
scikit-learn==1.4.0
tensorflow==2.15.0

# Domain analysis
dnspython==2.6.0
aiohttp==3.9.3
python-whois==0.8.0
tldextract==5.1.1

# Testing dependencies
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
